+++ Project/ProjectMono/ProjectMono.Service/VehicleModelService.cs
@@ -18,29 +18,52 @@ namespace ProjectMono.Service
         {
             this.repository = repository;
         }
+
+        //Create
         public async Task AddVehicleModelAsync(IVehicleModel entity)
         {
             await repository.AddVehicleModelAsync(entity);
         }
 
-        public async Task DeleteVehicleModelAsync(IVehicleModel entity)
+        //Delete
+        public async Task<Boolean> DeleteVehicleModelAsync(int id)
         {
-            await repository.DeleteVehicleModelAsync(entity);
+            var entity = await repository.GetVehicleModelAsync(id);
+            if(entity != null)
+            {
+                await repository.DeleteVehicleModelAsync(entity);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
 
+        //Read
+        //Gets a single VehicleModel by Id from repository
         public async Task<IVehicleModel> GetVehicleModelAsync(int id)
         {
             return await repository.GetVehicleModelAsync(id);
         }
-
+        //Gets sorted, flitered and paged VehicleModel list from repositroy
         public async Task<IPagedResult<IVehicleModel>> GetVehicleModelsAsync(ISortParameters sortParameters, IFilterParameters filterParameters, IPageParameters pageParameters, int? makeId)
         {
             return await repository.GetVehicleModelsAsync(sortParameters, filterParameters, pageParameters, makeId);
         }
 
-        public async Task UpdateVehicleModelAsync(IVehicleModel entity)
+        //Update
+        public async Task<Boolean> UpdateVehicleModelAsync(IVehicleModel entity)
         {
-            await repository.UpdateVehicleModelAsync(entity);
+            if (await repository.GetVehicleModelAsync(entity.Id) != null)
+            {
+                await repository.UpdateVehicleModelAsync(entity);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
     }
 }
